Improve useServerThemes in its own file
